+++ bambi/families/multivariate.py
@@ -5,7 +5,7 @@ import xarray as xr
 
 from bambi.families.family import Family
 from bambi.transformations import transformations_namespace
-from bambi.utils import extract_argument_names, get_aliased_name
+from bambi.utils import extract_argument_names, get_aliased_name, response_evaluate_new_data
 
 
 class MultivariateFamily(Family):
@@ -39,6 +39,23 @@ class Multinomial(MultivariateFamily):
         n = model.response_component.response_term.data.sum(1).astype(int)
         dont_reshape = ["n"]
         return super().posterior_predictive(model, posterior, n=n, dont_reshape=dont_reshape)
+
+    def log_likelihood(self, model, posterior, data, **kwargs):
+        if data is None:
+            y = model.response_component.response_term.data
+            trials = model.response_component.response_term.data.sum(1).astype(int)
+        else:
+            y = response_evaluate_new_data(model, data).astype(int)
+            trials = y.sum(1).astype(int)
+
+        # Prepend 'draw' and 'chain' dimensions
+        y = y[np.newaxis, np.newaxis, :]
+        trials = trials[np.newaxis, np.newaxis, :]
+
+        dont_reshape = ["n"]
+        return super().log_likelihood(
+            model, posterior, data=None, y=y, n=trials, dont_reshape=dont_reshape, **kwargs
+        )
 
     def get_coords(self, response):
         # For the moment, it always uses the first column as reference.