+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Interpet submodule now outputs informative messages when computing default values (#745) 
 * Bambi supports weighted responses (#761)
 * Bambi supports constrained responses (#764)
+* Implement `compute_log_likelihood()` method to compute the log likelihood on a model (#769)
 
 ### Maintenance and fixes